refactor(views): share the rule reordering between Rules and Dns
Both pages had the same drag-and-drop and move-up/down block, differing only in
whether it spliced `rules` or `dnsRules`, and both carried a <style scoped>
whose only content was an identical .mv-btn.

useReorder takes a getter and only splices. That is load-bearing rather than
stylistic: `rules` and `dnsRules` are readonly computeds over the array on the
config object, so an implementation that reassigned would log "Write operation
failed: computed value is readonly" and silently do nothing. A Ref<any[]>
parameter would have accepted those computeds at compile time and hidden it, so
the signature rules it out.

.mv-btn moves to app.css next to .kv-row, which already holds the other
in-card element classes. It is not folded into CardBtn: CardBtn's disabled
state means "request in flight" (opacity .5, wait cursor) where these mean "row
is already at an edge" (opacity .3, default cursor).

Dropping the scoped blocks takes .mv-btn's specificity from (0,2,0) to (0,1,0).
Checked in the built CSS: the only bare-button rules are under .segmented and
.tabs, neither of which is an ancestor here, and RuleCard — whose slot these
live in — has no <style> at all.
