+++ frontend/src/views/Dns.vue
@@ -151,6 +151,7 @@ import DnsRuleDrawer from '@/layouts/drawers/dns/DnsRuleDrawer.vue'
 import { Config } from '@/types/config'
 import { actionDnsRuleKeys, dnsRule } from '@/types/dns'
 import { FindDiff } from '@/plugins/utils'
+import { useReorder } from '@/plugins/useReorder'
 import { dnsColor } from '@/plugins/colors'
 import Btn from '@/components/ui/Btn.vue'
 import Ico from '@/components/ui/Ico.vue'
@@ -316,40 +317,5 @@ const confirmDelete = () => {
 }
 
 /* ---------------- ordering (drag & drop + move buttons) ---------------- */
-const draggedItemIndex = ref<number | null>(null)
-const onDragStart = (index: number) => { draggedItemIndex.value = index }
-const onDrop = (index: number) => {
-  if (draggedItemIndex.value !== null) {
-    const draggedItem = dnsRules.value[draggedItemIndex.value]
-    dnsRules.value.splice(draggedItemIndex.value, 1)
-    dnsRules.value.splice(index, 0, draggedItem)
-    draggedItemIndex.value = null
-  }
-}
-const moveRule = (index: number, dir: number) => {
-  const to = index + dir
-  if (to < 0 || to >= dnsRules.value.length) return
-  const item = dnsRules.value[index]
-  dnsRules.value.splice(index, 1)
-  dnsRules.value.splice(to, 0, item)
-}
+const { onDragStart, onDrop, move: moveRule } = useReorder(() => dnsRules.value)
 </script>
-
-<style scoped>
-.mv-btn {
-  flex: none;
-  width: 46px;
-  height: 44px;
-  border: none;
-  border-inline-start: 1px solid var(--line);
-  background: transparent;
-  cursor: pointer;
-  color: var(--text-2);
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  transition: all .14s;
-}
-.mv-btn:hover:not(:disabled) { background: var(--surface-3); color: var(--text); }
-.mv-btn:disabled { opacity: .3; cursor: default; }
-</style>
